+++ src/main/java/cart/controller/dto/response/ItemResponse.java
@@ -1,4 +1,4 @@
-package cart.controller.dto;
+package cart.controller.dto.response;
 
 import cart.service.dto.ItemDto;
